backend/requirements.txt
fastapi==0.103.0
setuptools==69.2.0
uvicorn==0.23.2
pydantic==1.10.5
python-dotenv==1.0.0
python-docx==1.1.0
docxtpl==0.16.7
dependency_injector==4.41.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
python-dateutil==2.8.2
psycopg2==2.9.9
jinja2==3.1.2
python-keycloak==3.7.0
alembic==1.12.1
Faker==20.1.0
pypdf==3.17.1
openai==1.3.8
dateparser==1.2.0
docx2python==2.8.0